fix map strategy_dict key to double in Robot.h

+++ C++/Robot.h
@@ -21,7 +21,7 @@ public:
     double y_;
     int todo_type;
     vector<tuple<double, double, int>> task_list;
-    map<int, double> strategy_dict;
+    map<double, double> strategy_dict;
     int last_assigned_time;
